cloudfunction/requirements.txt
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
cloudevents==1.11.0
deprecation==2.1.0
flask==3.1.2
functions-framework==3.9.2
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
opentelemetry-api==1.38.0
packaging==25.0
polars==1.34.0
polars-runtime-32==1.34.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
starlette==0.48.0
typing-extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvicorn-worker==0.4.0
watchdog==6.0.0
werkzeug==3.1.3
zipp==3.23.0
